define TableNamespace and pass to TableMapping (#25926)
as we start adding components to the data model, we will want to namespace the table mapping.
i.e. if you look up table by name "users", you get a different tablet depending on the component you're in.
and if you look up table by name "_modules", you get a different tablet depending on the component definition you're in.

This PR creates a new enum TableNamespace, although only `TableNamespace::Global` is used for now -- since it corresponds to all existing tables, and the namespacing is not persisted.

It also passes the namespace to all methods of `TableMapping` that need it. But it doesn't pass the namespace through in most places.

GitOrigin-RevId: 906ff095677693bf30be45f6a56058cad088e7b4