fix deleting tables with big search indexes (#25975)
when commiting a transaction, we look at all of the search indexes in the base snapshot. and we get their tablet by looking it up in the pending transaction snapshot, so we can determine whether that table was modified.

if you try to delete a table with a too-big search index, this is a problem: the index exists in the base snapshot but not in the pending transaction snapshot, so we throw an error before we can discover that the table was not modified.

fix the problem by looking up the index's table in the same snapshot where we found the index in the first place.

GitOrigin-RevId: f1dec3bc7beda8f7215b05f98e8c401e1ebd44d4