Account for all deletes when building a new segment (#25876)
Deletes must either delete documents in another segment or delete a document that was added in the revision stream that is being processed when building a new segment. This PR verifies this by tracking whether a document has a dangling deletion that should be accounted for by the end of the stream.

GitOrigin-RevId: 65ebc7973d05aafa3d9aba080735b1f03f792e56