Implement build_disk_index for fragmented text indexes (#25949)
The two significant sets of changes here are to add `Arc<dyn PersistenceReader>` to all of the flushers (most of the changes) and some refactoring in `incremental_index.rs` to return a set of paths (`TextSegmentPaths`) instead of using constants in the given directory. The first is required to allow text search to read revision pairs. The second aligns text search's index building with vector searches, making it easier to satisfy the API.

There's room for improvement with the index build API. TableIterator doesn't appear to be in timestamp order, so we're making some assumptions throughout the build index path that we only need descending document order if there are deletes. Not the cleanest way to do this.

GitOrigin-RevId: 36a55a6a8798ff5e19ed19c1f0466bd90c098a6c