Rename BackendApi to ApplicationApi and move it to application crate (#26013)
The Api needs to be moved to application as it is going to be moved to the sync worker. The name ApplicationApi is also better aligned with our current code structure where Application is the struct implementing all the logic and backend is a thin binary on top of it.

GitOrigin-RevId: 2fe3a5642de299353b1b3b5babf82e32eaebbfa1