fix SourceMap parsing (#26064)
instead of doing some custom parsing of the single field that we care about, let's use a library built for parsing source maps.
i didn't notice we already use this library in the node executor. this should avoid the errors we're getting with the null source content.

works as before for normal source maps. i don't know how to generate a null source map, but i checked the code and it's definitely handled.

<img width="1401" alt="Screenshot 2024-05-20 at 9 47 38 PM" src="https://github.com/get-convex/convex/assets/4319355/f2bd8e94-a0e3-49f2-bcf5-c33f4058c859">

GitOrigin-RevId: 81685983783b9f9bf53eddd26eb3cbdc9821679c