feat(components): Add enum to component metadata for distinguishing the root app from child components (#26097)
This moves to an enum so we don't have synthesize `name` and `args` for the app. We can let these two variants diverge more over time.

One interesting bit is that I left the `path: ComponentDefinitionPath` field alone, which means that the root app has definition path `""`, which makes sense as a relative path from `"convex/"`.

GitOrigin-RevId: b4d085cc7b4fd576557bb710298ae08b3e538b7c