Add a helper to spawn tokio tasks, use it for text/vector segments (#26094)
We're going to be spawning a lot, so there's a fair bit of repetitive code to create a channel, spawn a task, wait for future, send the result, then receive the result. It's not a ton of code, but we can save a bit with helpers.

The return types make this slightly annoying. If you do just `T` instaed of `Result<T>`, callers get an awkward double `Result<Result<T>>` which is more annoying to do deal with (e.g. `try_collect` gives you a `Vec<Result<T>>`). Since most callers will probably have futures that return results, and its much cleaner to unwrap the result in the method, I've gone that route. We could also add a second set of helpers that don't expect a result.

GitOrigin-RevId: a784af027340689da3c3f768df9cb5b279629678