Add flow for updating event format for log streams (#25993)
This is gated behind a flag, so should have no visible changes to existing users.
When the flag is turned on, users with existing log streams on the legacy format will see instructions to update to the new format, and new log streams will be created with the new format.

GitOrigin-RevId: 1c9e6ec7b6b63fbfef22ade3383469ba28fe4b7a