suspend actions timeout while preloading modules (#26132)
we suspend the actions timeout while fetching modules metadata, but not while fetching module source. we should do that.

we don't often suspend the actions timeout, because we can't do it for node actions, but i think this case warrants it, because node actions also don't count source package loading towards the timeout.

this shouldn't matter much, because the actions timeout is 10 minutes, but it does affect tests that are covering slow operations and so artificially set the actions timeout to a low value.

GitOrigin-RevId: 11628150eb6dd1d543f588f6883e0188c66f438e