refactor QueryPage cursor assertion (#26173)
move the assertion that pagination always has a cursor. now it is closer to the place where the constraint is guaranteed (or in the case of the bug we're tracking down, not guaranteed).

this just moves code, it does not change behavior so we should expected to see the error same as before.
i checked that there are no early returns or anything between the old assertion place and the new assertion place.

GitOrigin-RevId: 6856deaf5f1fa9d2a707ec99788250ea9f82f15f