error on pagination edge cases (#26190)
if you do the following in a UDF, you can cause a system error loop with an error message about the cursor not being set.

```
await ctx.db.query("messages").paginate({
      ...args.paginationOpts,
      maximumRowsRead: 0,
      // or maximumBytesRead: 0,
} as PaginationOptions);
```

fix this by catching the zero case early.

and also detect when we hit a pagination limit before setting the cursor. this should never happen, because in order to hit a pagination limit you need to have fetched some data, and if you fetched data then you must have set the cursor. if the scenario ever does happen, we can log the pagination error, which should help narrow down what's going on.

GitOrigin-RevId: 491bceae908bce0c6812378e3c160de3c303983d