make ActionPhase and ModuleModel component-aware (#26232)
actions don't understand components because they use `ComponentPath::root()` instead of the function's component path, and they look up resources by `ComponentId::Child` even if it's the root (it would be nice if this worked but it's used as BTreeMap keys which makes it tricky).

modules don't understand components because they store `_module_versions` in the wrong namespace, and sometimes they query and delete from the wrong namespace.

fix these holes

GitOrigin-RevId: 17ea9fedc1ba9c78e71280101ee766e787539dac