Short circuit scoring when multi segment text results are empty (#26202)
Talked with @emmaling27 - my understanding is filter terms are not included in the term dictionary and we don't need to do anything specific to delete them or account for deleted filter terms when scoring. Instead we rely on the search terms naturally excluding any corresponding filter terms.

The two places this shows up are:

1. During index builds we can't find deleted filter terms in the term dictionary
2. During search, it's theoretically possible that the `or` terms are empty, but the `and` terms are non-empty.

This PR treats the `or` / `search` terms as canonical, excluding filter terms from the deleted files when building and ignoring results with no `or` terms in the query path.

GitOrigin-RevId: 655385f5f04444532b067011a7612b36798949e3