[components] unique table numbers across components (#26245)
table numbers should stay unique across namespaces, even when the same system table is created in multiple namespaces. we do this by allowing tests to pick new table numbers, and making sure the uniqueness enforcement (which is not performance sensitive because it only happens when making a new table) looks across all namespaces.

GitOrigin-RevId: 0a7c0c34ed6274a8a1e4e6ad591ea8827737b749