Sync Worker: Always compute transition at latest timestamp (#26258)
I am converting sync worker to work with Usher, and if I keep the code as is, we will make multiple calls to latest_timestamp() per transition. One to determined the target timestamp and one to check it.

Making update_scheduled a Option<Timestamp> is generally a good idea, but we never actually leveraged it. So it is easier to revert for now. I have sent a separate RFC on how to finish the swing, but this provides little benefit until we have sync workers running on the edge.

GitOrigin-RevId: 8b40ebb2d5c83b50dd7ed8cac26f6da8b088cc29