move ConfigModel::apply contents to respective models (#26275)
`ConfigModel::apply` does a lot of things. As we add components to the system, some of these things will be scoped to component definitions, like `_modules` and `_schemas` changes, while others will be scoped to the components, like `_index` and `_cron_jobs`, and some will be scoped to be global, like `_auth`.

make these things easier to change scope of by doing a pure refactor that pulls each model's changes into a single method on the model in question. So all of the `_modules` changes are now in `ModuleModel::apply`

GitOrigin-RevId: 5a27dd69c6d5da7c13e19cd36234ea28527202a4