[dash] Use columns from enforced schema or shape in data page (#26244)
If there's an enforced schema, and we can determine the full set of fields from the schema (no `v.any()`), use those fields as the columns to show in the data page.

Otherwise, combine the fields from the shape + the fields from the currently loaded data to determine which columns to show.

GitOrigin-RevId: 6dadc022ab8826fcba0412193849502daf0f8117