Wrap mutable fragmented vector segments in a PreviousSegments struct (#26254)
I'll need this when we generalize the vector index writer. The writer needs to be able to delete points from previous segments and upload them, which is easier if the segments are wrapped in a struct.

This also makes vector match text exactly. We could further generalize the download/upload segments methods after this change as well.

GitOrigin-RevId: f4712bd779c87759746f924c603018557eaaa6c8