Fix base snapshot to permit using an index in the same transaction it's created in (#26337)
We had a mild bug where if a transaction created a new index, it wouldn't be able to query it. This is because...
1. `TransactionIndex` would use the latest `IndexRegistry` for checking that the desired index is enabled.
2. `TransactionIndex::range_no_deps` queries both the base `DatabaseIndexSnapshot` as well as its pending writes in `database_index_writes` and merges the results.
3. `DatabaseIndexSnapshot` has its own copy of the `IndexRegistry` from the _beginning_ of the transaction and throws an error if the queried index isn't enabled in that snapshot.

We previously suppressed the error in (3) for the `by_id` and `by_creation_time` index but not for other ones. This PR changes this special case to apply to all indexes that are not on the `_index` table, which is global and should always exist.

This issue came up in components work since we'd like to initialize the `_modules` table in a new component's namespace and then immediately put some rows into that table. `ModuleModel` uses the `by_path` index on this table to check for a previous version, and this query was failing at the base snapshot layer.

GitOrigin-RevId: 1747c0fc3fc56cb955c474d2029c7ab8f474abe4