Generalize logging text/vector segment/index metadata after a flush (#26326)
We're going to care about similar things for text / vector search and I'd like to generalize the remainder of the flushers. So I've created a combined set of metrics that can be used with both.

I ended up expanding the metrics slightly to include both the total number of documents as well as the number of non-deleted documents in each new segment and each index. Total number of documents gives us some idea of how big an index is (actual byte size would be nice too). Non-deleted documents lets us track how much overhead there is for each segment due to deletions.

I also removed logging the size of each segment after a flush. Now we only log the size of the new segment and the total size of the index across all segments. I'll have to fix a few graphs, but this seems more coherent.

GitOrigin-RevId: 1be03cb43651c98477ee7166c829527645da17ce