Change TestRuntime to be backed by Tokio (#23886)
The runtime-agnostic futures dream is over. There are plenty of places where we'd like to specialize for Tokio, and our generic `Runtime` trait prevents us from doing so.

This PR is the bare minimum to remove `TestRuntime`'s implementation but leaves the trait in place. We can eventually remove it later.

GitOrigin-RevId: 41c0479d57e34d72642227916cea11d4b1691b47