Partially revert 'Make sync worker state store authentication token' (#26382)
The original code validated the auth token when the Authenticate message was received and then checked its expiry time on every transition, mutation or action.

The new code only checks auth on every query, mutatoin or action. This means that the new code allow sending empty transition to the client even if their auth is invalid. This is technically correct from authorization stand point but results in visual blip on the clients with expired auth token. Partially revert the original PR as a stop gap measure. Will send a doc with proposal for the best way forward.

GitOrigin-RevId: 324d0cdf7765bb005d4895779de0401e783d16d6