Include Field in term metadata to support search using all fields (#26370)
This PR updates the deletion tracker and `segment_term_metadata` RPC to include `Field` data. It turns out that tantivy term ordinals are only unique to the term dictionary in an inverted index, which is specific to a field, so we need `Field` number to keep track of term deletions correctly.

* I had to move some structs and protos around to reduce redundancy, eventually these protos should be private.
* The naming is kind of a mess. We have a bunch of ways of representing term deletion metadata. I'm going to take a stab at cleaning up - suggestions are welcome.
* There's a follow up task to add proptests for the deletion tracker file format.

GitOrigin-RevId: 7f6b213d6389426dcf24a6b52193f1b980e82d89