Add a size_bytes_total field to fragmented text segment (#26371)
The compactor uses size to determine which segments should be compacted when. For vector segments this is relatively easy to estimate because all vectors are the same size and, we expect, the size of the filter fields is dwarfed by the size of the vectors.

We don't have any good heuristic for text segments. We do use one to estimate when to stop building an incremental segment during the initial backfill, but it's relatively inaccurate. For that case we don't care hugely if we're off by even an order of magnitude (if it's an over estimate at least).

One reason it feels reasonable to be off during the initial build is that we'll clean up overly small segments via compaction. For that work, compaction needs to know the actual size of the segment or something not too far off. It's pretty easy to obtain this size right after each segment is built so I've gone that route. We're not currently storing these segments anywhere so we shouldn't have any concerns about backfilling (like we would if we added this data to the vector metadata).

I have chosen not to update the segment size during deletions after the segment is first built. I assume that those deletions do not significantly impact the size of the data we store. We could recompute the size after each deletion, but doing so would require more code and we'd have to store the size on a per file basis for the mutable files, which seems like overkill.

GitOrigin-RevId: 5bf5503973063135d52fccbec015cf4ffca748f9