Add end_ts to ApplicationSubscription implementation. (#26395)
This is an RFC to illustrate how adding end_ts in ApplicationSubscription would look like. The tldr is that it doesn't affect behavior and only performance. Having end_ts allows optimizing calling extend_validity multiple times.

In practice, this optimization doesn't matter as long as the subscription worker can keep up but send this since there is nothing better than explaining trade-offs than code.

Proposal: Do not land this but open to clean up and land if @sujayakar thinks it is cleaner.
GitOrigin-RevId: f992a8e119880e6299643c753ec65435e85e1e92