Always pick recent creationTime in Function Runner (#26410)
Function runner uses the transaction being timestamp for initial creationTime. This means that if the backend is idle, the creationTime might be stale. To fix this, copy the logic from the backend where the creationTime also takes runtime.now() into account.

The backend already got the logic correct but next_creation_time is not serialized to protos, unless there are any writes. If there are any writes, the next_creation_time would be advanced correctly.

Alternative solution is for the backend to pick the next_creation_time and for Funrun to just use that. The upside is that variance in Funrun clock won't cause variance in creation time. The downside is that creationTime will be more stale if request takes longer to make it to Funrun or is a retry, etc. This is possible in large mutations that get retried due to OCC.

I am going with the simpler approach given we don't validate any invariants with creation time outside of it is >= transaction timestamp and is relevantly recent. It is also more aligned with our long term infrastructure goals where mutation execution can start without going via the backend first.

GitOrigin-RevId: 6634bf7e53d74ba04a8d063477c08bc1f5d14d9b