Change Application Api methods to use Identity instead of AuthentationToken (#26459)
I made the method signature for the http endpoints, but they force the sync worker to authenticate multiple times. This is technically fine since authentication should be instant as long as the auth provider metadata is cached. However, outside of being ugly it is also a big regression for misbehaving auth providers that do not set caching headers properly.

After some discussion, I have decided to optimize for the sync worker. The http api need to do an extra round trip between Backend and Usher for now, but this is noise in grand scheme of things. We can also easily optimize this later by forking the method or at least making auth an enum (identity or authentication token). However, not worth doing this for now.

GitOrigin-RevId: 029cf58b41dc874a149399e5ab9f543dd3e9ee14