Allow any Authorization header for HTTP actions (#26468)
HTTP actions should be allowed to set any valid header value for `Authorization`, not just the ones Convex uses for its other HTTP endpoint.

This was mostly fixed previously with `TryExtractIdentity`, where we computed the identity based on the `Authorization` header, allowing it to not correspond to a valid Convex identity, and use that if the function calls `auth.getUserIdentity`.

I didn't properly handle the `ExtractAuthenticationToken` failing (e.g. with `Authorization: a`) though. This is fixed and also has a regression test.

GitOrigin-RevId: 486af214c121a129d7a4d16da456906eab09284f