[Document Retention] Hash on `id` not `timestamp` and `id` (#26486)
The goal of this partitioning is that all documents with the same `id` are deleted in the same chunk. Except, we were hashing on the `id` and `timestamp`, so documents with the same `id` wouldn't always be in the same chunk. This is bad because we can delete more documents than intended.

Now, we will hash only based off of the document `id`. I added a test and confirmed that this failed with the old hashing method.

GitOrigin-RevId: d00d8692d70253762c9ec6762abe1e2e7934db2e