[REPL] write modules in separate method (#26617)
copy `ModuleModel::put` into a new method `ModuleModel::put_standalone` which has the same behavior.
only difference is the arguments are a little more restrictive, and there's a comment about the transaction and caching flow.

we do this so that we can (soon 🤞 )  remove the `put_module_source_into_db` from `ModuleModel::put`

GitOrigin-RevId: 7c4f79a3e59b83c2c3fb0c103963fbf6b1a27815