Add UdfCallbacks for calling from an IsolateClient back into it (#26680)
This PR adds support for queries calling into queries and mutations calling into queries and mutations.

It uses the same approach from the prototype: Pass a `UdfCallback` down into the `UdfEnvironment` that runs another function call within the same process. As @ldanilek had pointed out to me, this isn't actually that bad! When we move to `isolate2`, we'll be executing the cross component calls in the same process anyways, so it's not too bad to redirect them to the same function runner today.

To run a cross component call, we take out the current `Transaction` from the `UdfPhase`, pass it to the other call, get the result, put it back on our `UdfPhase`, and resume execution.

Since we don't have any form of subtransactions yet, there isn't any partial rollback. If a cross component call fails and the caller catches the error, its writes may still be "dirty" on the transaction.

GitOrigin-RevId: cdd2f87ed39233b86dc37539d9728f346208fefc