Fix schema short circuiting due to shape (#26342)
We had an issue a while ago where if you specified a schema like `defineTable({ _id: v.string(), foo: v.number() })`, we'd save the schema with `_id` as a string, which would violate internal expectations that `_id` has type `v.id(<table>)`.

We also realized that this meant that we were never hitting the optimization where we'd skip running schema validation on all existing data if the new schema was a subset of the shape of all the existing data (the shape always contained `_id` and `_creationTime`, schemas usually did not). This PR fixes this issue by stripping the top level system fields in the validator derived from the shape before doing the subset comparison.

We still haven't banned including `_id` and `_creationTime` in schemas -- we just quietly strip them out. We could do this now, but I don't think there's a huge benefit.

GitOrigin-RevId: 5d0fc0e466b6113141945c32f1ca3d52d0e46603