Remove search/vector sublimits, relax restrictions for adding pending indexes (#26722)
We were inconsistently counting search and vector indexes against the index limit. When we add a new index, they were counted against the limit. But when added in a database schema, they were not.

Now we always count both search and vector indexes against the global per table index limit. We no longer have a separate sublimit for any index type.

I also noticed that when an index was added, we would fail if we were over the global index limit, even if the new index was a backfilling version of an existing index. We also were counting both pending and enabled copies of any given index name against the limit. Since pending indexes are more or less invisible to the user and a pre-requisite for replacing or updating an index, they shouldn't count against the limit twice. We should also be able to add a new pending index for an existing enabled index even if the user is at the index limit so that the user can still update indexes when at the limit.

GitOrigin-RevId: dace7a0548dada60bf98b8af2c723bf700f56bd8