Fix term/docs statistics diff (#26810)
For multi segment search -
We need to track total number of deleted terms within all deleted documents as well as the number of deleted documents. Prior to this change we were counting the number of unique deleted terms as well as the number of documents. The mismatch in deleted terms count (unique vs non-unique) was caught by the memory index counting unique terms, resulting in a proptest failure.

I've added the proptest that caught this. It's more complicated than it strictly needs to be, but it's helpful in that it causes enough of a difference in score that it catches the bug even despite our score rounding.

I also refactored segment term metadata a bit. Before we'd pass through enough information to calculate the updated deletes per segment within Searcher. I've now made the RPC a bit simpler so that it just maps TermValues to TermOrdinals. That let's us pull the business logic of calculating the deletes up a level, so we can fix bugs in this area without having to change Searcher in the future.

GitOrigin-RevId: 811099082c672168ecefebff8bfb550a520c78a8