fix `next_system_table_number` on conflict  (#26844)
when picking a new system table number, try to pick one in the range allocated for system table numbers.
currently we take the max table number and add 1, but if there are any user tables, this picks a number which isn't allowed for system tables.
so instead we should have a candidate table number and keep incrementing it while there is a conflict with some existing table number.

GitOrigin-RevId: 6d14d1f2ba2516ea7bf435e9415216e6081d7e50