Only write to proto files when they change (#26850)
I believe that `cargo`'s internal mtime tracking is very simple: it just keeps the last mtime of a `cargo build` and reruns a build whenever a tracked file has an mtime greater than the last build timestamp.

This change makes it so that we'll get at most one spurious invalidation rather than continually recompiling `fivetran_common`:
1. `cargo build` runs w/o the proto file locally
2. the `build.rs` script downloads it and writes it with an mtime greater than (1)'s build timestamp.
3. the next `cargo build` recompiles `fivetran_common` but doesn't write the file to disk again, leaving the mtime as is.
4. subsequent `cargo build`s are clean since the proto file's mtime is before the build timestamp.

I'll stack the branch I was working on and validate that this fix works.

GitOrigin-RevId: eeeceb905e53a7a13e3dafa87aa494eccd25346f