use SourcePackageId instead of ModuleVersion as cache key (#26934)
in my bid to get rid of module versions, we can now start using source package id as the key for cache lookups.
this is more reliable because source packages don't disappear when the module gets updated.
and once we do this we can allow REPL lookups to go through the cache. without the package id change, we can't because REPL module version might conflict with a real module version, but the package ids won't conflict.

GitOrigin-RevId: a4429407817c8fe1e2eda998c3f0ac06bbb19551