make num_terms_deleted u64 (#27021)
This is a little safer because `num_terms_deleted` can be big now that it's not a unique count of terms deleted but a count of all the terms deleted.

If we sneak this in before any non-test instances are using deletion tracker files, then we should be ok to avoid a version bump.

GitOrigin-RevId: 7599f0b503874853676a96fa199c88eddd413a52