Track index intervals as we consume the documents instead of when we prefetch (#26853)
We were including the full interval in our readset when we fetched an index range. But since we often overfetch to filter out results later, this meant we had an overly broad readset (e.g. only selecting one document, but invalidating the query when a different document becomes second). Instead, track the interval when we consume the document.

GitOrigin-RevId: e18988a0710753f4e58500cddcf0dd6293d2e3a0