Handle changing term values due to replaced documents in the log (#27102)
When we iterate over the document log, we may encounter a document whose most recent previous revision does not match the version of that document that's in a previous text search segment. We need these to match because we update statistics based on the contents of the document. If the document terms we use to update statistics don't match the document in the previous segment, lots of things break.

To make this work, rather than taking the first previous revision of a document we find, we now look for the first previous revision of a document that's outside the timestamp bounds of the new segment. If the previous revision is within our new segment's bound, we skip processing it. As we continue backwards down the document log, we will then either find a pure add within our segment's time bounds (which we'll skip), or we eventually find a previous revision outside of our time bounds that we can use for the update.

GitOrigin-RevId: 7eb7a742fb77b5a85adcc6bea70f4c2b56adbc29