ModuleCache key on path, not module id (#27140)
make the module cache key be source_package_id and module path, which more closely identifies the stable source of truth in S3.

the advantage of this is we don't need to fetch module metadata for the things we're prefetching, which is necessary to map from path to module id. Just read the stuff out of the source package and stuff it in the cache.

GitOrigin-RevId: 504fbc72db2e1d1ed76c47ba1a926cf7781a6bd0