[components] namespace scheduled jobs (#27231)
get the scheduler working within components.
this is the basic functionality, where a component schedules its own function and that function runs.
a lot of the scaffolding is there for the rest of the features -- like the virtual table, cancellation, and cleanup worker, but we can address and test those later.

GitOrigin-RevId: 936820825bb6346aac8bf695c845342cb846cfc6