Add basic UI for snapshot import progress to the dashboard (#27272)
We currently show progress messages in the CLI, but if you disconnect or kill the CLI command you can no longer see the status of your import.

This is a minimal UI in the dashboard and put behind a flag. We probably want to slightly change the backend to make this sort of UI easier to build (e.g. split out the import summary from the confirmation message, potentially back this with a deployment audit log so we can know who performed it, potentially save the file name), but that's a more involved change.

We can also change the CLI to list the ongoing imports before starting a new one to prevent developers from inadvertently starting two overlapping imports as a separate change.

GitOrigin-RevId: 64106f597705b0fba1c4f226e62f850f87c560d2