Merge deletes correctly for text indexes (#27302)
This PR fixes a bug where we were not updating the deleted terms metadata when reprocessing deletes that happened in a race with a compaction. I added a `merge_deletes` method to the search index trait and had to thread through repeatable persistence to look up previous revisions for text search. It's a shame that we have to re-walk the document log and previous revisions in this race, but it also seems complicated to store the deletes and try to remap them to the right segments because we don't know how many compactions occur while we're writing the new segment to disk.

GitOrigin-RevId: e7df528990f8d36bd25b0be11a6b841e192e55ce