Remove `USE_MULTI_SEGMENT_SEARCH_QUERY` flag and associated code (#27077)
This flag is no longer very useful - the `BUILD_MULTI_SEGMENT_TEXT_INDEXES` flag controls whether we build or query multi segment text indexes. This does use the `search2` query path for both single and multisegment formats and removes the `Ranking` code which is unused on the multisegment query path.

GitOrigin-RevId: 272e7d55bd2413d2d44a9d9b52dd7b9e099144a8