remove a bunch of allow(dead_code) and change the rest to expect(dead_code) (#27383)
In the process of trying to clean up some `#[allow(dead_code)]` annotations, I found many that didn't even need the annotation in the first place.
I switched us to use `#[expect(dead_code)]` instead as this will warn if the annotation because unnecessary.
I had to make some changes to crate structure as some of the spurious warnings were caused by modules being redefined in a lib and bin; if code is shared, it should just be declared in the lib.

GitOrigin-RevId: 8dd3614ebf9556d4b927bc312f219b0a22a36f83