Allow pushing code when convex is a peer dependency. (#27422)
Allow pushing code when a package has convex as a peer dependency. This is useful when writing example apps in libraries: the repo might contain convex functions to push for testing but convex isn't a dependency of the project. It's possible but confusing to list a dependency as both a devDependency and a peerDependency. So just allow 'convex' to be in peerDependencies.

GitOrigin-RevId: d8afec97e917f408b9eca2b1b29ac4a87112df70