Remove custom process and console types (#27473)
Long ago, before 1.0 and before even Node.js actions, we suggested using separate global types for code in the convex directory. Adding Node.js actions made this completely infeasible, but it never worked very well because it behaved differently in editors than it did in `tsc`. We left this in to avoid changing behavior users with tsconfig.json files referencing it.

The presence of this file is currently breaking types in the open source JS repo so it's time to remove it. Even for developers still using the default tsconfig.json from before Novermber 28 2022 this shouldn't be a breaking change unless you don't have `@types/node` installed for your project; if so, install that and you should be good.

GitOrigin-RevId: e9b08d080059cbedc60e03aceeecf9a465028319