begin passing ComponentId through file storage (#27609)
whenever you're interacting with file storage, it's within the context of a component.
low-level APIs that operate on the table level are given a TableNamespace, while higher level interfaces use ComponentId. We could consider passing it around as ComponentPath instead, but I would rather not convert from ComponentId -> ComponentPath -> ComponentId in places like funrun callbacks. And ComponentId seems to work.

this PR passes the component down from a higher layer of code. It doesn't add the component to the protos or urls where it will eventually go.

The pattern for passing component id into `TaskExecutor` is the same pattern we use for `resources`, but it feels a little funny to copy that pattern.

GitOrigin-RevId: 210eaa19b8282d9c01a9da2971e13b44179585e7