Display functions that use paths starting with underscore in the dashboard (#27660)
If you have functions defined in something like `foo/_bar.ts` or `_foo.ts`, they'll get pushed and be runnable, but won't show up in the dashboard. This is because we
(1) filter out paths that begin with `_` in the system UDF to list modules.
(2) filter out path parts that begin with `_` in the dashboard `generateFileTree` code

I believe the goal here is to filter out system modules, but these aren't stored in the table we're querying, so these filters are only eliminating legitimate user modules that happen to start with `_`.

GitOrigin-RevId: 12ffc89661672e38219112b008bdac8903f059cc