Modify ApplicationApi subscription interface. (#27682)
Change the interface so each web socket creates a new subscription client. This lessens the requirement of the implementation since we can simply open a stream for each web socket and we don't have to recover from any errors.

We could change the interface and require more from the implementation - reopen stream on error as well as close it if there are no active clients. However, it is easier to not deal with this complexity for initial rollout. Again, we can always do this later.

From the backend perspective, it should be ok to move from one TPC socket per sync worker sessions to one stream per web socket sessions. We need to make sure to increase the #streams per grpc connection limit on the server.

GitOrigin-RevId: 03712521d1a8b5e6a7d999723a4b30f263bb85d9