Rename search indexes to text indexes (#27750)
This PR renames search indexes to text indexes when we mean full text search indexes. We have a bunch of generic code for building text and vector indexes where we use the term search index to mean both, so this should help clarify in the codebase what code is specific to text search. I didn't touch anything that's serialized because we should be more careful there about renames.

GitOrigin-RevId: 811120b188e43f02a60bafd0076f024a79eeb032