remove path from HTTP actions (#27782)
the path is always the same -- it's `http.js:default` and it's in the root component. if we end up allowing different possible paths, that will probably be configured in the routing layer; it won't be determined by the request. so we can stop passing this path down.

while i'm here, i noticed that `ComponentPath::root()` is being used in several places it shouldn't be. to help audit these, i added `TODO()` and `test_user()` constructors like we have on `ComponentId`.

GitOrigin-RevId: f51adaa1ee4c424e118180097d1107f4292f271d