Make scheduled mutations observe itself in progress (#27801)
Mark scheduled jobs as in-progress while they are executing. Given that mutations commit atomically, unlike actions, this state can only be observed by the mutation itself.

This property seems make the state machine more symmetric and complete, which is useful for libraries such as Cronvex that build on top of scheduled functions.

GitOrigin-RevId: dab4d0a66704b17be1ee7d69182f3c44e21be3d7