[components] use DeveloperDocumentId instead of InternalId (#27829)
i used to think storing InternalIds for references between system tables would be good, because it's the minimum information needed to identify a document if you already know the system table it's supposed to be in.

however, i'm changing my mind: i don't like InternalIds at all, because they require us to use table numbers (and I want to reduce usage of table numbers) and they make it hard to convert from _id field to the InternalId on the client. In particular I want a system UDF that lists components and gets their IDs, but that's actually difficult because components have an `_id` and a `ComponentId`, and the system UDF can't convert between them easily.

therefore I want to change references within _components and _component_definitions to use DeveloperDocumentId instead of InternalId everywhere. This should be equivalent because system table numbers are stable, and it allows us to treat the ids as just strings. We should be able to do this without a migration because all apps using components are internal and we can deactivate them.

GitOrigin-RevId: 881fa5ac363f30b62b0712dda38d1212677913fe