[components] dashboard only show tables in current component (#27828)
currently, the dashboard lists the tables from all components, regardless of which component is selected in the UI. if the table isn't in the current component, it has a weird asterix next to it because it doesn't appear in the current schema, and it doesn't have any data in it.

to fix, we should pass the component id into the `/api/shapes2` endpoint and filter to just that component.

GitOrigin-RevId: b9edc24c6b5d5dee51a2bf5a08f69632e04b2c44